backend/apps/apiv1/views/users.py
from rest_framework.permissions import AllowAny
from rest_framework.views import APIView
from rest_framework_simplejwt.views import TokenObtainPairView
from rest_framework.response import Response
from django.contrib.auth import authenticate, logout

class SigninView(TokenObtainPairView):
    permission_classes = (AllowAny,)

class LogoutView(APIView):
    permission_classes = (AllowAny,)

    def post(self, request, *args, **kwargs):
        logout(request)
        return Response(status=204)