requirements.txt
appnope==0.1.4
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==2.0.12
decorator==5.1.1
Django==3.2.25
django-cors-headers==3.10.1
django-environ==0.11.2
django-webpack-loader==1.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.4.0
environ==1.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.56.3
httplib2==0.22.0
idna==3.10
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
oauth2client==4.1.3
oauthlib==3.2.2
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
protobuf==3.19.6
psycopg2==2.9.8
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.14.0
PyJWT==1.7.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
traitlets==4.3.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
zipp==3.6.0
